mysite/requirements.txt
django
djangorestframework
environs
django-cors-headers
django-filter